fixed dpo_better type error.

+++ dpo_better.py
@@ -16,7 +16,6 @@ from sft import SFT_DATASET
 MODEL_NAME = "Qwen/Qwen2.5-0.5B"
 COUNTDOWN_DATASET = "Jiayi-Pan/Countdown-Tasks-3to4"
 DPO_DATASET = "dpo_dataset_merged.json"
-#DPO_DATASET = "dpo_dataset_small.json"
 SFT_PATH = "models/sft/BEST_epochs_6-batch_4-lr_1e-05-seed_42-curr_type_none-scheduler_True-static_False-repeat_epochs_None.pt"
 device = 'cuda' if torch.cuda.is_available() else 'cpu'
 print(f"Using device: {device}")
@@ -75,7 +74,7 @@ def main(args):
     test_dataset = load_dataset(SFT_DATASET, split='test')
     test_dataloader = torch.utils.data.DataLoader(test_dataset, batch_size=args.batch_size)
 
-    prev_indices = np.array([])
+    prev_indices = np.array([], dtype=int)
     prev_losses = None
     best_val_loss = float('inf')
     for epoch in range(args.num_epochs):
